+++ src/ska_mid_dish_steering_control/constants.py
@@ -1,17 +1,18 @@
 """Common enumerated types and other constants used in the SCU package."""
 
 from enum import Enum, IntEnum
-from importlib import metadata
+from importlib.metadata import version
 from pathlib import Path
 from typing import Final
 
 from platformdirs import user_cache_dir
 
+__version__: Final = version("ska-mid-dish-steering-control")
+
 # Constants
 USER_CACHE_DIR: Final = Path(
     user_cache_dir(appauthor="SKAO", appname="ska-mid-dish-steering-control")
 )
-PACKAGE_VERSION: Final = metadata.version("ska-mid-dish-steering-control")
 SUBSCRIPTION_RATE_MS: Final = 100
 
 
@@ -109,3 +110,13 @@ class ResultCode(IntEnum):
 
 # Type aliases
 CmdReturn = tuple[ResultCode, str, list[int | None] | None]
+JSONData = (  # Type hint for any JSON-encodable data
+    None
+    | bool
+    | int
+    | float
+    | str
+    | list["JSONData"]  # A list can contain more JSON-encodable data
+    | dict[str, "JSONData"]  # A dict must have str keys and JSON-encodable data
+    | tuple["JSONData", ...]  # A tuple can contain more JSON-encodable data
+)