+++ CHANGELOG.rst
@@ -1,9 +1,13 @@
+Changelog
+---------
+
 Unreleased
-----------
+^^^^^^^^^^
 
 - API changes:
     - WOM-479: Removed redundant ``get_enum_strings()`` method. Use ``get_attribute_data_type()`` instead.
 - Non-breaking changes:
+    - WOM-386: Added new ``StaticPointingModel`` class for the import/export of a global static pointing model from/to a JSON file.
     - WOM-479: Updated ``get_attribute_data_type()`` to fix exceptions in DiSQ.
     - WOM-484: Updated ``subscribe()`` to subscribe to all input nodes in one OPCUA call.
     - WOM-446, WOM-464: Added missing commands to the ``Command`` enum.
@@ -15,7 +19,7 @@ Unreleased
 
 
 Older history
--------------
+^^^^^^^^^^^^^
 
 The source files of this project were migrated from the `ska-mid-disq 
 <https://gitlab.com/ska-telescope/ska-mid-disq>`_ project on 10 Sept 2024,